refactor(agent-engine): confine tool batches to transition
