+++ crates/agent-engine/src/runtime/mod.rs
@@ -27,9 +27,9 @@ mod response_guardrails;
 mod steering_queue;
 mod submission_handlers;
 mod tool_authorization;
+mod tool_batch;
 mod tool_effect_lifecycle;
 mod tool_execution;
-mod tool_orchestrator;
 mod tool_packages;
 mod tool_policy;
 mod tool_presentation;
