+++ crates/agent-engine/src/runtime/transition/tests.rs
@@ -23,6 +23,8 @@ use std::{
 use tempfile::TempDir;
 use tokio_util::sync::CancellationToken;
 
+mod tool_batch;
+
 async fn maybe_compact_context_for_request<E, F>(
     state: &mut RuntimeLoopState,
     emit: &mut E,
